More robust test of date for different timezones
For different timezones, dates can actually shift over to the next day. So let's only test for the format of the date. Minutes and seconds should still be the same for other timezones, so let's keep this check. Note that for timezone having an offset from UTC which is not a full hour, this might still fail. We'll address this if the problem actually arises.

Side note: I'm always amazed at how simple time looks, but how complex it really is ...